Refactor DDS example and tools to be able to run with/without DDS
