--- fairmq/deployment/CMakeLists.txt
@@ -1,60 +1,0 @@
- ################################################################################
- #    Copyright (C) 2014 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH    #
- #                                                                              #
- #              This software is distributed under the terms of the             #
- #         GNU Lesser General Public Licence version 3 (LGPL) version 3,        #
- #                  copied verbatim in the file "LICENSE"                       #
- ################################################################################
-
-Set(INCLUDE_DIRECTORIES
-  ${CMAKE_SOURCE_DIR}/fairmq
-  ${CMAKE_SOURCE_DIR}/fairmq/zeromq
-  ${CMAKE_SOURCE_DIR}/fairmq/nanomsg
-  ${CMAKE_SOURCE_DIR}/fairmq/devices
-  ${CMAKE_SOURCE_DIR}/fairmq/tools
-  ${CMAKE_SOURCE_DIR}/fairmq/options
-  ${CMAKE_SOURCE_DIR}/fairmq/deployment
-  ${CMAKE_CURRENT_BINARY_DIR}
-)
-
-Set(SYSTEM_INCLUDE_DIRECTORIES
-  ${SYSTEM_INCLUDE_DIRECTORIES}
-  ${Boost_INCLUDE_DIR}
-  ${DDS_INCLUDE_DIR}
-  ${ZMQ_INCLUDE_DIR}
-)
-
-Include_Directories(${INCLUDE_DIRECTORIES})
-Include_Directories(SYSTEM ${SYSTEM_INCLUDE_DIRECTORIES})
-
-Set(LINK_DIRECTORIES
-  ${LINK_DIRECTORIES}
-  ${Boost_LIBRARY_DIRS}
-  ${DDS_LIBRARY_DIR}
-)
-
-Link_Directories(${LINK_DIRECTORIES})
-
-Install(FILES FairMQDDSTools.h DESTINATION include)
-
-Set(Exe_Names
-  ${Exe_Names}
-  fairmq-dds-command-ui
-)
-
-Set(Exe_Source
-  ${Exe_Source}
-  runDDSCommandUI.cxx
-)
-
-list(LENGTH Exe_Names _length)
-math(EXPR _length ${_length}-1)
-
-ForEach(_file RANGE 0 ${_length})
-  list(GET Exe_Names ${_file} _name)
-  list(GET Exe_Source ${_file} _src)
-  set(EXE_NAME ${_name})
-  set(SRCS ${_src})
-  set(DEPENDENCIES FairMQ dds_intercom_lib)
-  GENERATE_EXECUTABLE()
-EndForEach(_file RANGE 0 ${_length})
